se coloco la tarea android
