--- BuildConfig.java
@@ -1,13 +1,0 @@
-/**
- * Automatically generated file. DO NOT MODIFY
- */
-package com.autonoma.login;
-
-public final class BuildConfig {
-  public static final boolean DEBUG = Boolean.parseBoolean("true");
-  public static final String APPLICATION_ID = "com.autonoma.login";
-  public static final String BUILD_TYPE = "debug";
-  public static final String FLAVOR = "";
-  public static final int VERSION_CODE = 1;
-  public static final String VERSION_NAME = "1.0";
-}
