--- README.md
@@ -1,2 +1,0 @@
-# Tarea-andorid
-mi primera tarea
