+++ app/src/main/res/values/strings.xml
@@ -1,0 +1,7 @@
+<resources>
+    <string name="app_name">Login</string>
+    <string name="login">Login</string>
+    <string name="usuario">Usuario</string>
+    <string name="clave">Clave</string>
+    <string name="validar">Validar</string>
+</resources>
